apps/api/pyproject.toml
[project]
name = "pathfinder-api"
version = "0.1.0a1"
description = "VEuPathDB Strategy Builder Chatbot API"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "kani[openai,anthropic,google]>=1.2.0",
    "qdrant-client>=1.12.0",
    "python-multipart>=0.0.16",
    "structlog>=24.4.0",
    "pyyaml>=6.0.2",
    "tenacity>=9.0.0",
    "json5>=0.9.0",
    "rapidfuzz>=3.9.0",
    "optuna>=4.0.0",
    "redis[hiredis]>=5.2.0",
    "pyjwt>=2.12.0",
    "slowapi>=0.1.9",
    "pathfinder-shared",
    "furo[dev]>=2025.12.19",
    "sphinx-design[dev]>=0.7.0",
    "sphinx-copybutton[dev]>=0.5.2",
    "sphinxcontrib-mermaid[dev]>=2.0.1",
    "sphinx-inline-tabs[dev]>=2025.12.21.14",
    "sphinx[dev]>=8",
]

[tool.uv]
constraint-dependencies = [
    "pyasn1>=0.6.3",  # CVE-2026-30922
]

[tool.uv.sources]
pathfinder-shared = { path = "../../packages/shared-py", editable = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "respx>=0.21.0",
    "pre-commit>=3.7.0",
    "ruff>=0.7.0",
    "mypy>=1.19.0",
    "types-pyyaml>=6.0.0",
    "vulture>=2.10",
    "sphinx>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/veupath_chatbot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["src/veupath_chatbot/tests"]
markers = [
    "live_wdk: tests that make live HTTP calls to VEuPathDB WDK endpoints (require network)",
]

[tool.coverage.run]
source = ["src/veupath_chatbot"]
omit = ["*/tests/*"]

[tool.ruff]
line-length = 88
target-version = "py314"
src = ["src", "alembic"]

[tool.ruff.lint]
select = [
    "E", "F", "I", "N", "W", "UP", "B", "C4", "SIM",  # existing
    "A",      # builtin-shadowing (list = [], type = "foo")
    "ASYNC",  # async anti-patterns (missing await, blocking in async)
    "BLE",    # blind except (bare `except:` or `except Exception:`)
    "C90",    # McCabe complexity
    "DTZ",    # datetime without timezone
    "EM",     # string literals in exceptions
    "ERA",    # commented-out code
    "FBT",    # boolean trap (boolean positional args)
    "ICN",    # import conventions
    "ISC",    # implicit string concatenation
    "LOG",    # logging anti-patterns
    "PIE",    # misc linting
    "PT",     # pytest style enforcement
    "PTH",    # pathlib over os.path
    "RET",    # unnecessary return/else patterns
    "RSE",    # raise improvements
    "RUF",    # ruff-specific rules
    "S",      # bandit security (SQL injection, hardcoded passwords, etc.)
    "T20",    # print() statements (use structlog)
    "TCH",    # TYPE_CHECKING imports
    "TID",    # tidy imports
    "TRY",    # exception handling anti-patterns
    "PERF",   # performance anti-patterns
    "FURB",   # modernization
    "PLC",    # pylint convention
    "PLE",    # pylint error
    "PLR",    # pylint refactor
    "PLW",    # pylint warning
]
ignore = [
    # FastAPI path params and Pydantic field names intentionally use mixedCase
    # to match the public HTTP contract (e.g. `siteId`, `planSessionId`).
    # Long prompt/template strings intentionally exceed line-length.
    "E501",
    "N803",
    "N815",
    # Pydantic models with self-referential fields need quoted forward refs
    # even on Python 3.14 (PEP 649) because Pydantic's metaclass evaluates
    # annotations eagerly at class construction time.
    "UP037",
    # We don't use TYPE_CHECKING imports.
    "TC001", "TC002", "TC003",
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pylint]
max-args = 6
max-returns = 3
max-branches = 10

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = [
    "S101",     # assert is what tests do
    "S105",     # hardcoded passwords in test fixtures are fine
    "S106",     # hardcoded password args in test fixtures are fine
    "PLR2004",  # magic values in assertions are expected
    "FBT",      # boolean args in test helpers are fine
    "S311",     # deterministic seeds for shuffling is fine
]
"src/veupath_chatbot/ai/prompts/**" = [
    "E501",  # prompt strings are long by nature
]

[tool.ruff.lint.isort]
known-first-party = ["veupath_chatbot"]

[tool.mypy]
python_version = "3.14"
strict = true
mypy_path = "stubs"
warn_return_any = true
warn_unused_ignores = true
disable_error_code = ["import-untyped", "untyped-decorator"]
exclude = ["scripts/", "docs/", "src/veupath_chatbot/tests/"]

[[tool.mypy.overrides]]
module = "veupath_chatbot.tests.*"
disable_error_code = [
    "no-untyped-def",
    "arg-type",
    "assignment",
    "return-value",
    "type-arg",
    "var-annotated",
]

[tool.vulture]
paths = ["src/veupath_chatbot", "vulture_whitelist.py"]
exclude = ["tests/", "*/conftest.py"]
min_confidence = 80

[dependency-groups]
dev = [
    "fakeredis>=2.34.1",
    "import-linter>=2.0",
    "pip-audit>=2.7.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "respx>=0.22.0",
    "ruff>=0.15.0",
    "testcontainers",
]


[tool.pyright]
pythonVersion = "3.14"
pythonPlatform = "Linux"

# ---------------------------------------------------------------------------
# Import Linter — backend architecture enforcement
# ---------------------------------------------------------------------------
[tool.importlinter]
root_packages = ["veupath_chatbot"]
include_external_packages = true

[[tool.importlinter.contracts]]
name = "Domain layer is pure (no I/O, no other layers)"
type = "forbidden"
source_modules = ["veupath_chatbot.domain"]
forbidden_modules = [
    "veupath_chatbot.integrations",
    "veupath_chatbot.services",
    "veupath_chatbot.transport",
    "veupath_chatbot.persistence",
    "veupath_chatbot.ai",
    "httpx",
    "sqlalchemy",
    "asyncpg",
    "qdrant_client",
    "redis",
    "fastapi",
]

[[tool.importlinter.contracts]]
name = "Transport never imports integrations or persistence directly"
type = "forbidden"
source_modules = ["veupath_chatbot.transport"]
forbidden_modules = [
    "veupath_chatbot.integrations",
    "veupath_chatbot.persistence",
]

[[tool.importlinter.contracts]]
name = "AI tools never import integrations or persistence directly"
type = "forbidden"
source_modules = ["veupath_chatbot.ai.tools"]
forbidden_modules = [
    "veupath_chatbot.integrations",
    "veupath_chatbot.persistence",
]

[[tool.importlinter.contracts]]
name = "Services never import transport or AI"
type = "forbidden"
source_modules = ["veupath_chatbot.services"]
forbidden_modules = [
    "veupath_chatbot.transport",
    "veupath_chatbot.ai",
]

[[tool.importlinter.contracts]]
name = "Integrations never import services, transport, or AI"
type = "forbidden"
source_modules = ["veupath_chatbot.integrations"]
forbidden_modules = [
    "veupath_chatbot.services",
    "veupath_chatbot.transport",
    "veupath_chatbot.ai",
]

[[tool.importlinter.contracts]]
name = "Persistence never imports services, transport, AI, or integrations"
type = "forbidden"
source_modules = ["veupath_chatbot.persistence"]
forbidden_modules = [
    "veupath_chatbot.services",
    "veupath_chatbot.transport",
    "veupath_chatbot.ai",
    "veupath_chatbot.integrations",
]

